+++ src/CDTypeLexer.m
@@ -41,7 +41,7 @@ static NSString *CDTypeLexerStateName(CDTypeLexerState state)
 
 - (void)setState:(CDTypeLexerState)newState;
 {
-    if (debug) NSLog(@"CDTypeLexer - changing state from %lu (%@) to %lu (%@)", _state, CDTypeLexerStateName(_state), newState, CDTypeLexerStateName(newState));
+    if (debug) NSLog(@"CDTypeLexer - changing state from %lu (%@) to %lu (%@)", (unsigned long)_state, CDTypeLexerStateName(_state), (unsigned long)newState, CDTypeLexerStateName(newState));
     _state = newState;
 }
 
@@ -58,7 +58,7 @@ static NSString *CDTypeLexerStateName(CDTypeLexerState state)
     _lexText = nil;
 
     if ([_scanner isAtEnd]) {
-        if (_shouldShowLexing)                       NSLog(@"%s [state=%lu], token = TK_EOS", sel_getName(_cmd), _state);
+        if (_shouldShowLexing)                       NSLog(@"%s [state=%lu], token = TK_EOS", sel_getName(_cmd), (unsigned long)_state);
         return TK_EOS;
     }
 
@@ -66,23 +66,23 @@ static NSString *CDTypeLexerStateName(CDTypeLexerState state)
         // Skip whitespace, scan '<', ',', '>'.  Everything else is lumped together as a string.
         [_scanner setCharactersToBeSkipped:[NSCharacterSet whitespaceCharacterSet]];
         if ([_scanner scanString:@"<" intoString:NULL]) {
-            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = %d '%c'", sel_getName(_cmd), _state, '<', '<');
+            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = %d '%c'", sel_getName(_cmd), (unsigned long)_state, '<', '<');
             return '<';
         }
 
         if ([_scanner scanString:@">" intoString:NULL]) {
-            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = %d '%c'", sel_getName(_cmd), _state, '>', '>');
+            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = %d '%c'", sel_getName(_cmd), (unsigned long)_state, '>', '>');
             return '>';
         }
 
         if ([_scanner scanString:@"," intoString:NULL]) {
-            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = %d '%c'", sel_getName(_cmd), _state, ',', ',');
+            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = %d '%c'", sel_getName(_cmd), (unsigned long)_state, ',', ',');
             return ',';
         }
 
         if ([_scanner my_scanCharactersFromSet:[NSScanner cdTemplateTypeCharacterSet] intoString:&str]) {
             _lexText = str;
-            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = TK_TEMPLATE_TYPE (%@)", sel_getName(_cmd), _state, _lexText);
+            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = TK_TEMPLATE_TYPE (%@)", sel_getName(_cmd), (unsigned long)_state, _lexText);
             return TK_TEMPLATE_TYPE;
         }
 
@@ -95,7 +95,7 @@ static NSString *CDTypeLexerStateName(CDTypeLexerState state)
 
         if ([_scanner scanIdentifierIntoString:&identifier]) {
             _lexText = identifier;
-            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = TK_IDENTIFIER (%@)", sel_getName(_cmd), _state, _lexText);
+            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = TK_IDENTIFIER (%@)", sel_getName(_cmd), (unsigned long)_state, _lexText);
             _state = CDTypeLexerState_Normal;
             return TK_IDENTIFIER;
         }
@@ -109,23 +109,23 @@ static NSString *CDTypeLexerStateName(CDTypeLexerState state)
                 _lexText = @"";
 
             [_scanner scanString:@"\"" intoString:NULL];
-            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = TK_QUOTED_STRING (%@)", sel_getName(_cmd), _state, _lexText);
+            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = TK_QUOTED_STRING (%@)", sel_getName(_cmd), (unsigned long)_state, _lexText);
             return TK_QUOTED_STRING;
         }
 
         if ([_scanner my_scanCharactersFromSet:[NSCharacterSet decimalDigitCharacterSet] intoString:&str]) {
             _lexText = str;
-            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = TK_NUMBER (%@)", sel_getName(_cmd), _state, _lexText);
+            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = TK_NUMBER (%@)", sel_getName(_cmd), (unsigned long)_state, _lexText);
             return TK_NUMBER;
         }
 
         if ([_scanner scanCharacter:&ch]) {
-            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = %d '%c'", sel_getName(_cmd), _state, ch, ch);
+            if (_shouldShowLexing)                   NSLog(@"%s [state=%lu], token = %d '%c'", sel_getName(_cmd), (unsigned long)_state, ch, ch);
             return ch;
         }
     }
 
-    if (_shouldShowLexing)                           NSLog(@"%s [state=%lu], token = TK_EOS", sel_getName(_cmd), _state);
+    if (_shouldShowLexing)                           NSLog(@"%s [state=%lu], token = TK_EOS", sel_getName(_cmd), (unsigned long)_state);
 
     return TK_EOS;
 }
